Remove unnecessary complexity from the backend GET /work endpoint
A bit of backstory for this commit, since it doesn't look like much: the
original intention was to query all the work details in one sql query
instead of the multiple queries in the fetch_work_details function, but
it quickly got very complex and I realized that the work listing doesn't
need the details anyway. Now the works endpoint is ~20x faster (based on
anecdotal evidence from the debugging logs), yay.